refactor: split oversized hotspots for maintainability
Break up three large files into focused, cohesive modules:

ChatViewModel.swift (2076 → 841 lines):
- ChatSessionError.swift: SessionErrorCategory, SessionError types
- ChatViewModel+Attachments.swift: file/image attachment handling
- ChatViewModel+MessageHandling.swift: daemon message dispatch

AppDelegate.swift (1553 → 806 lines):
- AppDelegate+MenuBar.swift: status menu, skills, menu actions
- AppDelegate+Sessions.swift: session lifecycle, escalation
- AppDelegate+Notifications.swift: notification setup, delegate
- AppDelegate+BundleHandling.swift: .vellumapp file handling

session.ts (1465 → 1236 lines):
- session-process.ts: drainQueue, processMessage extraction
- session-tool-setup.ts: tool definitions, executor callback
